Refactor and Design DIR
