+++ .env
@@ -1,3 +1,3 @@
-# NODE_ENV = 'production'
-NODE_ENV = 'development'
+NODE_ENV = 'production'
+# NODE_ENV = 'development'
 PORT = 3009
