+++ .gitignore
@@ -1,5 +1,10 @@
 npm
 npm-debug.log
 node_modules/
-/.vscode
+.vscode/
 .env
+
+## Keep Certain Build Files
+build/*.gz
+build/*.html
+build/*.js
