--- client/config.js
@@ -1,4 +1,0 @@
-
-module.exports = {
-    awsReview: 'http://localhost:3009'
-}
