--- client/src/components/blank.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-let Blank = (props) => (
-        <div className="blank"><h4>This page is under construction.</h4></div>
-)
-
-export default Blank
