+++ server/database/mySchema.sql
@@ -1,0 +1,15 @@
+CREATE TABLE reviews (
+  id INTEGER NOT NULL AUTO_INCREMENT DEFAULT NULL,
+  item_id INTEGER NOT NULL DEFAULT NOT NULL,
+  title VARCHAR(50) NOT NULL,
+  pros VARCHAR(100) NOT NULL,
+  cons VARCHAR(100) NOT NULL,
+  body VARCHAR(300) NOT NULL,
+  verified CHAR(1) NOT NULL,
+  date TIMESTAMP NOT NULL,
+  eggs INTEGER NOT NULL DEFAULT 0,
+  author VARCHAR(30) NOT NULL DEFAULT 'Anonymous',
+  helpful INTEGER NULL DEFAULT 0,
+  not_helpful INTEGER NOT NULL DEFAULT 0,
+  PRIMARY KEY (id)
+);
