+++ src/config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  awsReview: 'http://localhost:3009'
+};
