+++ src/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Reviews</title>
+    <link rel="stylesheet" href="style.css" />
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css?family=Roboto:300,400"
+    />
+  </head>
+  <body>
+    <div id="reviewApp"></div>
+    <script type="text/javascript" src="bundle.js"></script>
+  </body>
+</html>
