+++ webpack.config.js
@@ -1,16 +1,17 @@
-var path = require('path');
-var SRC_DIR = path.join(__dirname, '/client/src');
-var DIST_DIR = path.join(__dirname, '/client/dist');
-var CompressionPlugin = require('compression-webpack-plugin');
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+const CompressionPlugin = require('compression-webpack-plugin');
 
 module.exports = {
   mode: process.env.NODE_ENV || 'development',
   entry: {
-    app: ['babel-polyfill', `${SRC_DIR}/index.jsx`]
+    app: ['babel-polyfill', `./src/index.jsx`]
   },
   output: {
+    path: path.resolve(__dirname, 'build'),
     filename: 'bundle.js',
-    path: DIST_DIR
+    chunkFilename: '[id].js',
+    publicPath: ''
   },
   module: {
     rules: [
@@ -24,5 +25,12 @@ module.exports = {
       }
     ]
   },
-  plugins: [new CompressionPlugin()]
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: __dirname + '/src/index.html',
+      inject: 'body',
+      filename: 'index.html'
+    }),
+    new CompressionPlugin()
+  ]
 };
